requirements.txt
absl-py==1.0.0
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.0.3
antlr4-python3-runtime==4.8
anyio==3.3.2
appdirs==1.4.4
argon2-cffi==21.1.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
black==21.4b2
bleach==1.5.0
blinker==1.4
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2022.5.18.1
cffi==1.14.6
charset-normalizer==2.0.12
click
cloudpickle==2.0.0
cryptography==36.0.2
cycler==0.10.0
Cython==0.29.24
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
deprecation==2.1.0
detectron2
dpcpp-cpp-rt==2022.1.0
entrypoints==0.3
flatbuffers==2.0
fonttools
frozenlist
future==0.18.2
fvcore==0.1.5.post20211023
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py
html5lib==0.9999999
hydra-core==1.1.1
idna==3.3
imageio==2.9.0
imgaug==0.4.0
importlib-metadata
importlib-resources==5.2.2
intel-cmplr-lib-rt==2022.1.0
intel-cmplr-lic-rt==2022.1.0
intel-opencl-rt==2022.1.0
intel-openmp==2022.1.0
iopath==0.1.8
ipykernel
ipython
ipython-genutils
ipywidgets
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.5
jupyter-core==4.8.1
jupyter-packaging==0.10.6
jupyter-server==1.11.0
jupyterlab==3.1.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
kaolin
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver
libclang
Markdown
MarkupSafe
matplotlib
matplotlib-inline
mistune
mkl==2022.1.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mmcv==1.5.0
multidict==6.0.2
mypy-extensions==0.4.3
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
nn-distance==0.0.0
nose==1.3.7
notebook==6.4.4
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
omegaconf==2.1.1
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pip==21.2.4
plyfile==0.7.4
portalocker==2.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.19.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.21
pydot==1.4.2
Pygments==2.10.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.1
pyu2f==0.1.5
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.3.0
regex==2021.8.28
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.8
scikit-image==0.18.3
scikit-learn==1.0
scipy==1.7.3
Send2Trash==1.8.0
setuptools==61.2.0
Shapely==1.7.1
six==1.16.0
sniffio==1.2.0
tabulate==0.8.9
tbb==2021.6.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.1
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==2.2.0
tifffile==2021.8.30
toml==0.10.2
tomlkit==0.7.2

tornado==6.1
tqdm==4.64.0
traitlets==5.1.0
typing_extensions==4.2.0
urllib3==1.26.9
usd-core==22.5.post1
wcwidth==0.2.5
websocket-client==1.2.1
Werkzeug==2.1.2
wheel==0.37.1
widgetsnbextension==3.5.1
wrapt==1.14.1
yacs==0.1.8
yapf==0.31.0
yarl==1.7.2
zipp==3.8.0
